refactor: 보드 상태를 보여주는 로직을 BoardGame에게 위임
